ml-models/requirements.txt
# ── Core ──────────────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
joblib>=1.3.0

# ── Deep Learning ─────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0

# ── Anomaly Detection ─────────────────────────────────────────
# Isolation Forest via scikit-learn (above)
# LSTM Autoencoder via PyTorch (above)

# ── Time-Series Forecasting ──────────────────────────────────
# LSTM Forecaster via PyTorch (above)

# ── Root Cause Classification ─────────────────────────────────
xgboost>=2.0.0
imbalanced-learn>=0.11.0

# ── Log Clustering ────────────────────────────────────────────
sentence-transformers>=2.2.0
hdbscan>=0.8.33
umap-learn>=0.5.4

# ── NLP Report Generation (Phi-3-mini quantised) ─────────────
transformers>=4.40.0
accelerate>=0.27.0
bitsandbytes>=0.43.0

# ── Data & Evaluation ────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0.0
tqdm>=4.65.0
requests>=2.31.0

# ── Serving ──────────────────────────────────────────────────
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# ── Experiment Tracking (optional but recommended) ───────────
mlflow>=2.10.0
